.agents/skills/skill-creator/pyproject.toml
[project]
name = "skill-creator"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0.3",
]
